AVRO-444. Fix warnings. Contributed by Bruce Mitchener.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/avro/trunk@918833 13f79535-47bb-0310-9956-ffa450edef68

+++ lang/c/src/io.c
@@ -167,7 +167,7 @@ avro_read_file(struct avro_file_reader_t *reader, void *buf, int64_t len)
 		return 0;
 	}
 
-	if (needed > sizeof(reader->buffer)) {
+	if (needed > (int64_t) sizeof(reader->buffer)) {
 		if (bytes_available(reader) > 0) {
 			memcpy(p, reader->cur, bytes_available(reader));
 			p += bytes_available(reader);
